+++ src/FTabs/FTabsController/types.ts
@@ -1,0 +1,4 @@
+export interface FTabsControllerProps {
+	tabKey: string;
+	setTabKey: (key: string) => void;
+}