+++ src/FTabs/FTabsHeader/types.ts
@@ -1,0 +1,12 @@
+import { CSSProperties } from "react";
+import { FButtonProps } from "../../FButton/types";
+import { FTabsControllerProps } from "../FTabsController/types";
+import { FTabsPanelProps } from "../FTabsPanel/types";
+
+export interface FTabsHeaderProps {
+	controller: FTabsControllerProps;
+	panelProps: FTabsPanelProps[];
+	style?: CSSProperties;
+	className?: string;
+	tabButtonProps?: FButtonProps;
+}