+++ src/FTabs/FTabsPanel/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+import { FTabsPanelProps } from "./types";
+
+export const FTabsPanel = (props: FTabsPanelProps) => {
+	const { children } = props;
+	return <>{children}</>;
+};