+++ src/FTabs/FTabsPanel/types.ts
@@ -1,0 +1,7 @@
+import { ReactNode } from "react";
+
+export interface FTabsPanelProps {
+	label: string;
+	tabKey: string;
+	children?: ReactNode;
+}